+++ src/splunk_as/__init__.py
@@ -177,7 +177,7 @@ from .validators import (
     validate_url,
 )
 
-__version__ = "1.1.3"
+__version__ = "1.1.4"
 __all__ = [
     # Version
     "__version__",